Revamp the build system to generate header guards
Header guards (in the modern sense - guards that forbid you from
including a header more than once) are now generated by an awk script.

There are two directives that may appear in a header.  The first is the
"provide" directive:

    //provide foo.h

This is expanded to #ifdef rubbish that prevents the header from being
included more than once.  This should be the file's name unless it is
one of several mutually exclusive headers (e.g. foo-oldway.h, foo-newway.h).

The second is the "require" directive:

    //require bar.h

This is expanded to #ifndef rubbish that gives a clean error if a header
that must be required before this one has not been included.

This is a much better way to manage headers than to use traditional
header guards and recursive includes, as these typically end up with
some common headers being included dozens or hundreds of times.